move server/utils to server/support/utils
