+++ generators/server/index.mjs
@@ -20,4 +20,4 @@
 export { default } from './generator.mjs';
 export { serverFiles as files } from './files.mjs';
 export * from './needles.mjs';
-export * from './utils.mjs';
+export * from './support/utils.mjs';
